+++ NEWS.md
@@ -3,6 +3,8 @@ wallace 2.1.3
 - Fixed NULL error in poccs_selectOccs Rmd code
 - Updated for new ENMeval v2.0.5 functionality
 - Replaced deprecated `ENMeval::getcheckerboard1()` and `ENMeval::getcheckerboard2()` with `ENMeval::getcheckerboard()`
+- Replaced `raster::getdata` with `geodata` functions
+- Now using WorldClim v2.1 env data
 
 wallace 2.1.2
 =============